moving code library state to separate file
